# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample API for xendit backend test. It contains API endpoints to create a new ride, retrieve a list of rides or retrieve a single ride.",
    "version": "1.0.0",
    "title": "Xendit Rides API"
  },
  "host": "localhost:8010",
  "basePath": "/",
  "tags": [
    {
      "name": "Rides",
      "description": "Everything about Rides"
    },
    {
      "name": "Health Check",
      "description": "Checking the API status"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/rides": {
      "post": {
        "tags": [
          "Rides"
        ],
        "summary": "Create a new ride",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Ride object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation"
          }
        }
      },
      "get": {
        "tags": [
          "Rides"
        ],
        "summary": "Get All Rides",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "$ref": "#/definitions/Rides"
            }
          }
        }
      }
    },
    "/rides/{id}": {
      "get": {
        "tags": [
          "Rides"
        ],
        "summary": "Find a Ride by ID",
        "description": "Returns a single ride",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of ride",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health Check"
        ],
        "summary": "check API status",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Healthy"
          }
        }
      }
    }
  },
  "definitions": {
    "Ride": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "start_lat": {
          "type": "integer",
          "format": "int64"
        },
        "start_long": {
          "type": "integer",
          "format": "int64"
        },
        "end_lat": {
          "type": "integer",
          "format": "int64"
        },
        "end_long": {
          "type": "integer",
          "format": "int64"
        },
        "rider_name": {
          "type": "string",
          "example": "Joko Suryo"
        },
        "driver_name": {
          "type": "string",
          "example": "Mpu Tantular"
        },
        "driver_vehicle": {
          "type": "string",
          "example": "Kencana Mas"
        }
      }
    },
    "Rides": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Ride"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}